// typescript-sdk/src/server/types/tracer.ts
export interface TypedValueChatMessages {
  type: "chat_messages";
  value: ChatMessage[];
}

interface TypedValueText {
  type: "text";
  value: string;
}

interface TypedValueRaw {
  type: "raw";
  value: string;
}

type JSONSerializable =
  | string
  | number
  | boolean
  | null
  | Record<string, any>
  | any[];

interface TypedValueJson {
  type: "json";
  value: JSONSerializable;
}

type Money = {
  currency: string;
  amount: number;
};

export interface GuardrailResult {
  status: "processed" | "skipped" | "error";
  passed: boolean;
  score?: number | null;
  details?: string | null;
  cost?: Money | null;
}

interface TypedValueGuardrailResult {
  type: "guardrail_result";
  value: GuardrailResult;
}

export type SpanInputOutput =
  | TypedValueText
  | TypedValueChatMessages
  | TypedValueGuardrailResult
  | TypedValueJson
  | TypedValueRaw;

export interface ErrorCapture {
  has_error: true;
  message: string;
  stacktrace: string[];
}

interface SpanMetrics {
  prompt_tokens?: number | null;
  completion_tokens?: number | null;
  tokens_estimated?: boolean | null;
  cost?: number | null;
}

interface SpanParams {
  temperature?: number;
  stream?: boolean;
  functions?: Record<string, any>[];
  tools?: Record<string, any>[];
  tool_choice?: string;
}

interface SpanTimestamps {
  started_at: number;
  first_token_at?: number | null;
  finished_at: number;
}

export type SpanTypes =
  | "span"
  | "llm"
  | "chain"
  | "tool"
  | "agent"
  | "rag"
  | "guardrail"
  | "unknown";

export interface BaseSpan {
  span_id: string;
  parent_id?: string | null;
  trace_id: string;
  type: SpanTypes;
  name?: string | null;
  input?: SpanInputOutput | null;
  outputs: SpanInputOutput[];
  error?: ErrorCapture | null;
  timestamps: SpanTimestamps;
  metrics?: SpanMetrics | null;
}

export interface LLMSpan extends BaseSpan {
  type: "llm";
  vendor: string;
  model: string;
  params: SpanParams;
}

export interface RAGChunk {
  document_id?: string | null;
  chunk_id?: string | null;
  content: string | Record<string, any> | any[];
}

export interface RAGSpan extends BaseSpan {
  type: "rag";
  contexts: RAGChunk[];
}

export type Span = LLMSpan | RAGSpan | BaseSpan;
// ...
export type TraceInput = {
  value: string;
  embeddings?: { model: string; embeddings: number[] };
  satisfaction_score?: number;
};

export type TraceOutput = {
  value: string;
  embeddings?: { model: string; embeddings: number[] };
};

export type Trace = {
  trace_id: string;
  project_id: string;
  metadata: {
    thread_id?: string;
    user_id?: string;
    customer_id?: string;
    labels?: string[];
    topic_id?: string;
    subtopic_id?: string;
  };
  timestamps: { started_at: number; inserted_at: number; updated_at: number };
  input: TraceInput;
  output?: TraceOutput;
  metrics: {
    first_token_ms?: number | null;
    total_time_ms?: number | null;
    prompt_tokens?: number | null;
    completion_tokens?: number | null;
    total_cost?: number | null;
    tokens_estimated?: boolean | null;
  };
  error?: ErrorCapture | null;
  indexing_md5s?: string[];
};